requirements.txt
streamlit
pytesseract
pillow
openpyxl
numpy
